+++ CHANGES.md
@@ -19,7 +19,9 @@ Beta Releases
    * Removed `affectedByLighting` property from `Polygon`, `EllipsoidPrimitive`, `RectangularPyramidSensorVolume`, `CustomSensorVolume`, and `ComplexConicSensorVolume`.  
    * Removed `DistanceIntervalMaterial`.  This was not documented.
    * `Matrix2.getElementIndex`, `Matrix3.getElementIndex`, and `Matrix4.getElementIndex` functions have had their parameters swapped and now take row first and column second.  This is consistent with other class constants, such as Matrix2.COLUMN1ROW2.
+   * `CentralBody.showSkyAtmosphere` now defaults to `true`.  This was already the default when using the Cesium widget.
 * Improved lighting in Columbus view and on polygons, ellipsoids, and sensors.
+* Improved atmosphere Columbus view transition.
 * Added `Water` material.  See the Materials Sandbox example.
 * Fixed jitter artifacts with billboards and polylines.
 * Added `SkyBox` to draw stars.  Added `CesiumWidget.showSkyBox` and `CesiumViewerWidget.showSkyBox`.